+++ src/fidesops/migrations/versions/07014ff34eb2_add_mariadb.py
@@ -16,7 +16,17 @@ depends_on = None
 
 
 def upgrade():
-    op.execute("alter type connectiontype add value 'mariadb'")
+    op.execute("alter type connectiontype rename to connectiontype_old")
+    op.execute(
+        "create type connectiontype as enum('postgres', 'mongodb', 'mysql', 'https', 'snowflake', 'redshift', 'mssql', 'mariadb')"
+    )
+    op.execute(
+        (
+            "alter table connectionconfig alter column connection_type type connectiontype using "
+            "connection_type::text::connectiontype"
+        )
+    )
+    op.execute("drop type connectiontype_old")
 
 
 def downgrade():